Address PR #198 review: reduce duplication, move fixtures to conftest
- Replace _get_aizynthfinder_cls() with top-level try-except import
- Extract _resolve_config() and _run_finder() helpers shared by both
  run_az and run_az_with_img, so config fallback logic is consistent
- Move test fixtures from test_az.py to conftest.py for reuse
